Jungi (#54)
* mygather and gather categorySearch method

* commit (#8)

* Create README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Haha (#9)

* commit

* commit

* commit

* Kjs (#11)

* commit

* commit

* commit

* commit

* commit

* Kjs (#14)

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* Kjs (#15)

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* gatherList and myGatherList and signUp Page

* Kjs (#22)

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* init location modal setting

* Jungi (#19)

* mygather and gather categorySearch method

* commit (#8)

* Create README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Haha (#9)

* commit

* commit

* commit

* Kjs (#11)

* commit

* commit

* commit

* commit

* commit

* Kjs (#14)

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* Kjs (#15)

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

---------


* Kjs (#22)

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* change dto and entity chatting and chattingMessage to GatherComment and GatherCommentMessage and mygatherlist and gatherlist logics

* change duplicate mapper method

* change duplicate mapper method

* change static SQL to Dynamic SQl

* change interface in interface

* Kjs (#34)

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* commit

* change gather directory

* change myGatherList view and pageCount

* add view Toggle function myGatherList and gatherList View

* add list image

* user entity userDto point field chagne and userList and MyuserList

---------

+++ src/main/java/com/example/batisproject/service/user/impl/UserServiceImpl.java
@@ -93,7 +93,7 @@ public class UserServiceImpl implements UserService {
 
         User target = userMapper.existsByEmail(userDTO.getUsername());
 
-        int totalPoint = target.getPoint() + userDTO.getPoint();
+        Long totalPoint = target.getPoint() + userDTO.getPoint();
         target.setPoint(totalPoint);
 
         int result = userMapper.updateUser(target);
